Prevent staging in sandbox

+++ src/Manifest.php
@@ -212,35 +212,38 @@ class Manifest
      */
     protected static function freshConfiguration($project)
     {
-        static::write(array_filter([
-            'id'           => $project['id'],
-            'name'         => $project['name'],
-            'environments' => [
-                'production' => array_filter([
-                    'memory'     => 1024,
-                    'cli-memory' => 512,
-                    'runtime'    => 'php-8.1:al2',
-                    'build'      => [
-                        'COMPOSER_MIRROR_PATH_REPOS=1 composer install --no-dev',
-                        'php artisan event:cache',
-                        file_exists(Path::current().'/webpack.mix.js')
-                            ? 'npm ci && npm run prod && rm -rf node_modules'
-                            : 'npm ci && npm run build && rm -rf node_modules',
-                    ],
-                ]),
-                'staging' => array_filter([
-                    'memory'     => 1024,
-                    'cli-memory' => 512,
-                    'runtime'    => 'php-8.1:al2',
-                    'build'      => [
-                        'COMPOSER_MIRROR_PATH_REPOS=1 composer install',
-                        'php artisan event:cache',
-                        file_exists(Path::current().'/webpack.mix.js')
-                            ? 'npm ci && npm run dev && rm -rf node_modules'
-                            : 'npm ci && npm run build && rm -rf node_modules',
-                    ],
-                ]),
+        $environments['production'] = array_filter([
+            'memory' => 1024,
+            'cli-memory' => 512,
+            'runtime' => 'php-8.1:al2',
+            'build' => [
+                'COMPOSER_MIRROR_PATH_REPOS=1 composer install --no-dev',
+                'php artisan event:cache',
+                file_exists(Path::current().'/webpack.mix.js')
+                    ? 'npm ci && npm run prod && rm -rf node_modules'
+                    : 'npm ci && npm run build && rm -rf node_modules',
             ],
+        ]);
+
+        if (! $project['is_sandbox']) {
+            $environments['staging'] = array_filter([
+                'memory' => 1024,
+                'cli-memory' => 512,
+                'runtime' => 'php-8.1:al2',
+                'build' => [
+                    'COMPOSER_MIRROR_PATH_REPOS=1 composer install',
+                    'php artisan event:cache',
+                    file_exists(Path::current().'/webpack.mix.js')
+                        ? 'npm ci && npm run dev && rm -rf node_modules'
+                        : 'npm ci && npm run build && rm -rf node_modules',
+                ],
+            ]);
+        }
+
+        static::write(array_filter([
+            'id' => $project['id'],
+            'name' => $project['name'],
+            'environments' => $environments,
         ]));
     }
